+++ crates/sdmx_json/src/data.rs
@@ -1,5 +1,5 @@
 use crate::primitives::{
-	Action, Annotation, DataType, Link, LocalizedText, MetaSingleReceiver, NumberOrString,
+	Action, Annotation, DataType, Link, LocalizedText, Meta, NumberOrString, SdmxMessage,
 	SdmxValue, StatusMessage,
 };
 use serde::{Deserialize, Serialize};
@@ -12,14 +12,29 @@ use std::str::FromStr;
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default)]
 pub struct DataMessage {
 	#[serde(skip_serializing_if = "Option::is_none")]
-	pub meta: Option<MetaSingleReceiver>,
+	pub meta: Option<Meta>,
 	#[serde(skip_serializing_if = "Option::is_none")]
 	pub data: Option<Data>,
 	#[serde(skip_serializing_if = "Option::is_none")]
-	pub error: Option<StatusMessage>,
+	pub errors: Option<Vec<StatusMessage>>,
 	#[serde(skip_serializing_if = "Option::is_none")]
 	#[serde(flatten)]
 	pub other: Option<HashMap<String, Value>>,
+}
+
+impl SdmxMessage for DataMessage {
+	type Data = Data;
+	fn meta(&self) -> Option<&Meta> {
+		self.meta.as_ref()
+	}
+
+	fn data(&self) -> Option<&Self::Data> {
+		self.data.as_ref()
+	}
+
+	fn errors(&self) -> Option<&Vec<StatusMessage>> {
+		self.errors.as_ref()
+	}
 }
 
 impl<'a> TryFrom<&'a [u8]> for DataMessage {